various fixes: publish photos checked by default. can't delete content packs from the home screen. can only delete content packs from the catalog screen that you have installed. sending messages on delete and download complete.
