+++ app/res/menu/activity_catalog.xml
@@ -1,0 +1,8 @@
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:id="@+id/menu_about"
+        android:title="@string/menu_about"
+        android:orderInCategory="4"
+        android:showAsAction="never" />
+
+</menu>
